requirements.txt
cohere==1.3.10
Flask==3.0.0
lyricsgenius==3.0.1
python-dotenv==1.0.0
requests==2.28.0
